remotion_one/pyproject.toml
[project]
name = "remotion-one"
version = "0.1.0"
description = "Article to explainer video tool powered by Remotion and LLM"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "claude-code-sdk>=0.0.20",
    "edge-tts>=7.2.1",
    "aiofiles>=24.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]

[project.scripts]
remotion-one = "remotion_one.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/remotion_one"]
